Venus_framework/requirement.txt
pillow>=9.0.0
requests>=2.28.0
pyyaml>=6.0
